[#17] refact: FortunePage UI 분리
