+++ App/Donmani/Feature/Fortune/Component/FortunePage.swift
@@ -6,8 +6,7 @@
 //
 
 import SwiftUI
-import Kingfisher
-import DesignSystem
+import Core
 import Domain
 
 struct FortunePage: View {
@@ -57,7 +56,17 @@ struct FortunePage: View {
 
             HStack(spacing: itemSpacing) {
                 ForEach(visibleFortunes, id: \.offset) { _, fortune in
-                    page(fortune: fortune)
+                    FortunePageItem(
+                        fortune: fortune,
+                        selectedDay: selectedDay,
+                        referenceToday: referenceToday,
+                        referenceYesterday: referenceYesterday,
+                        hasTodayRecord: hasTodayRecord,
+                        hasYesterdayRecord: hasYesterdayRecord,
+                        isFlipped: isFortuneCardFlipped(fortune.day),
+                        touchFortuneCardAction: touchFortuneCardAction,
+                        touchRecordAction: touchRecordAction
+                    )
                         .frame(width: cardWidth)
                 }
             }
@@ -101,86 +110,6 @@ struct FortunePage: View {
             dragOffset = 0
         }
         .frame(maxWidth: .infinity)
-    }
-
-    @ViewBuilder
-    private func page(fortune: Fortune) -> some View {
-        let isFlipped = isFortuneCardFlipped(fortune.day)
-        let isSelected = fortune.day == selectedDay
-        let recordButtonLabel = recordButtonLabel(for: fortune.day)
-
-        VStack(spacing: 16) {
-            FortuneFlipCard(
-                fortune: fortune,
-                isFlipped: isFlipped
-            )
-            .frame(height: 313)
-            .clipShape(
-                RoundedRectangle(
-                    cornerRadius: .s5,
-                    style: .continuous
-                )
-            )
-            .contentShape(
-                RoundedRectangle(
-                    cornerRadius: .s5,
-                    style: .continuous
-                )
-            )
-            .onTapGesture {
-                guard isSelected else {
-                    return
-                }
-                withAnimation(.easeInOut(duration: 0.36)) {
-                    touchFortuneCardAction(fortune.day)
-                }
-            }
-            .allowsHitTesting(isSelected)
-
-            VStack(alignment: .leading, spacing: 0) {
-                DText(
-                    fortune.day.fortuneDate,
-                    style: .b2,
-                    weight: .regular,
-                    color: Color.white
-                )
-                DText(
-                    fortune.subtitle,
-                    style: .h2,
-                    weight: .bold,
-                    color: Color.white
-                )
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-
-            if let recordButtonLabel {
-                Button {
-                    guard isSelected else {
-                        return
-                    }
-                    touchRecordAction(fortune.day)
-                } label: {
-                    ZStack {
-                        RoundedRectangle(
-                            cornerRadius: .s5,
-                            style: .continuous
-                        )
-                        .fill(ColorPalette.Primary.deepBlue20)
-                        DText(
-                            recordButtonLabel,
-                            style: .h3,
-                            weight: .bold,
-                            color: .white
-                        )
-                    }
-                }
-                .frame(height: 58)
-                .disabled(!isSelected)
-                .allowsHitTesting(isSelected)
-            }
-            
-            Spacer()
-        }
     }
 
     private enum Direction {
@@ -241,127 +170,6 @@ struct FortunePage: View {
             dragOffset = 0
             isTransitioning = false
         }
-    }
-
-    private func recordButtonLabel(for day: Day) -> String? {
-        if day == referenceToday, !hasTodayRecord {
-            return "오늘 소비 기록하기"
-        }
-        if day == referenceYesterday, !hasYesterdayRecord {
-            return "어제 소비 기록하기"
-        }
-        return nil
-    }
-}
-
-private struct FortuneFlipCard: View, Animatable {
-    let fortune: Fortune
-    var progress: CGFloat
-
-    init(
-        fortune: Fortune,
-        isFlipped: Bool
-    ) {
-        self.fortune = fortune
-        self.progress = isFlipped ? 1 : 0
-    }
-
-    var animatableData: CGFloat {
-        get { progress }
-        set { progress = newValue }
-    }
-
-    var body: some View {
-        ZStack {
-            if progress < 0.5 {
-                frontCardFace
-                    .scaleEffect(
-                        x: max(0.001, 1 - (progress * 2)),
-                        y: 1,
-                        anchor: .center
-                    )
-            } else {
-                cardContent
-                    .scaleEffect(
-                        x: max(0.001, (progress - 0.5) * 2),
-                        y: 1,
-                        anchor: .center
-                    )
-            }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-
-    @ViewBuilder
-    private var frontCardFace: some View {
-        if let imageURL = URL(string: fortune.imageUrl), !fortune.imageUrl.isEmpty {
-            KFImage(imageURL)
-                .placeholder {
-                    cardBackground(ColorPalette.Semantic.dailyFortuneBackground)
-                }
-                .cancelOnDisappear(false)
-                .resizable()
-                .scaledToFill()
-                .frame(maxWidth: .infinity, maxHeight: .infinity)
-                .clipped()
-        } else {
-            cardBackground(ColorPalette.Semantic.dailyFortuneBackground)
-        }
-    }
-
-    private func cardBackground(_ color: Color) -> some View {
-        RoundedRectangle(
-            cornerRadius: .s3,
-            style: .continuous
-        )
-        .fill(color)
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-
-    private var cardContent: some View {
-        VStack(spacing: 16) {
-            HStack(spacing: 8) {
-                DImage(DImageAsset.dailyFortune)
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(width: 68, height: 68)
-                VStack(
-                    alignment: .leading,
-                    spacing: 4
-                ) {
-                    DText("토비 요정이 전해주는", style: .b2, weight: .medium, color: .fromHex("#806AEB"))
-                        .multilineTextAlignment(.leading)
-                    DText(fortune.day.fortuneDate, style: .h3, weight: .bold, color: .fromHex("#04091E"))
-                        .multilineTextAlignment(.leading)
-                }
-                Spacer()
-            }
-            DText(fortune.content, style: .b1, weight: .regular, color: .fromHex("#04091E"))
-                .lineSpacing(8)
-            HStack {
-                DText("⭐️ \(fortune.item)", style: .b3, weight: .medium, color: .fromHex("#FFFFFF"))
-                    .kerning(-0.5)
-                    .padding(.vertical, 8)
-                    .padding(.horizontal, 8)
-                    .background {
-                        Capsule()
-                            .fill(Color.fromHex("#6045E6"))
-                    }
-                Spacer()
-            }
-            Spacer(minLength: 0)
-        }
-        .padding(20)
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .background {
-            cardBackground(ColorPalette.Semantic.dailyFortuneBackground)
-        }
-    }
-}
-
-private extension Array {
-    subscript(safe index: Int) -> Element? {
-        indices.contains(index) ? self[index] : nil
     }
 }
